add additional error handling to project routes

+++ routes/projectRoutes.js
@@ -13,6 +13,15 @@ routes.get("/api/projects", async (req, res, next) => {
 });
 
 routes.post("/api/projects", async (req, res, next) => {
+  const { body } = req;
+
+  if (!body.name || !body.description) {
+    next({
+        status: 400,
+        message: "A name and description must be provided"
+    })
+  }
+
   try {
     const project = await db.insert(req.body);
     res.status(200).json(project);
@@ -23,8 +32,24 @@ routes.post("/api/projects", async (req, res, next) => {
 
 routes.put("/api/projects/:id", async (req, res, next) => {
   const { id } = req.params;
+
+  if (!id) {
+    next({
+      status: 400,
+      message: "Please provide an id"
+    });
+  }
+
   try {
     const project = await db.update(id, req.body);
+
+    if (!project) {
+      next({
+        status: 404,
+        message: `Project with id ${id} could not be found`
+      });
+    }
+
     res.status(201).json(project);
   } catch (e) {
     next({
@@ -36,11 +61,26 @@ routes.put("/api/projects/:id", async (req, res, next) => {
 
 routes.delete("/api/projects/:id", async (req, res, next) => {
   const { id } = req.params;
+
+  if (!id) {
+    next({
+      status: 400,
+      message: "Please provide an id"
+    });
+  }
+
   try {
     const deletedProject = await db.remove(id);
+
+    if (!deletedProject) {
+      next({
+        status: 404,
+        message: `Project with id ${id} could not be found`
+      });
+    }
+
     res.status(201).json(deletedProject);
   } catch (e) {
-      console.log(e)
     next({
       status: 501,
       message: `Could not delete record ${id}`
